+++ NEWS.md
@@ -2,6 +2,8 @@
 
 ### NEW FEATURES
 
+* New function 'extended.pgls' for evaluating within-species trends in a phylogenetic context
+* New dataset 'pupfish.ws'
 * New function `physignal.eigen` for evaluating the dimensions of phylogenetic signal in multivariate data
 * Added an option to `gm.prcomp` to allow the phylogeny to be scaled to unit length
 * Added `groups.first` argument to `gm.measurement.error`, to optionally include groups in ANOVA table.